requirements.txt
charset-normalizer==3.4.2
coloredlogs==15.0.1
filelock==3.13.1
flatbuffers==25.2.10
fsspec==2024.6.1
humanfriendly==10.0
imbalanced-learn==0.13.0
Jinja2==3.1.4
joblib==1.5.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
onnx==1.18.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
packaging==25.0
pillow==11.0.0
protobuf==6.31.1
pyreadline3==3.5.4
scikit-learn==1.6.1
scipy==1.16.0
sklearn-compat==0.1.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchvision==0.22.0+cu126
typing_extensions==4.12.2
